Change decimal datatypes to float Create stored procedure for inserting into all tables at once Create view that selects from all tables give FDMS user EXECUTE permissions

+++ Database/Server/Altitude_Parameters.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [FDMS].[Altitude_Parameters]
 (
 	[Telemetry_ID] BIGINT NOT NULL PRIMARY KEY, 
-    [Altitude] DECIMAL NOT NULL, 
-    [Pitch] DECIMAL NOT NULL, 
-    [Bank] DECIMAL NOT NULL, 
+    [Altitude] FLOAT NOT NULL, 
+    [Pitch] FLOAT NOT NULL, 
+    [Bank] FLOAT NOT NULL, 
     CONSTRAINT [FK_Altitude_Parameters_To_Telemetry_Records] FOREIGN KEY ([Telemetry_ID]) REFERENCES [FDMS].[Telemetry_Records]([Telemetry_ID])
 )
